+++ packages/widgetbook_cli/lib/src/analyzer/collectors/widget_name_collector.dart
@@ -9,7 +9,7 @@ class WidgetNameCollector
   @override
   bool shouldCollect(ClassDeclaration node) {
     // Skip private classes
-    if (node.name.toString().startsWith('_')) return false;
+    if (node.namePart.toString().startsWith('_')) return false;
 
     // Skip classes that have the ignore comment
     final firstToken = node.firstTokenAfterCommentAndMetadata;
@@ -21,7 +21,7 @@ class WidgetNameCollector
 
   @override
   String mapNode(ClassDeclaration node) {
-    return node.name.toString();
+    return node.namePart.toString();
   }
 }