+++ packages/widgetbook_cli/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 
 - **BREAKING**: Set minimum SDK version to 3.9.0. ([#1803](https://github.com/widgetbook/widgetbook/pull/1803))
+- **REFACTOR**: Use `analyzer` 10.x. ([#1804](https://github.com/widgetbook/widgetbook/pull/1804))
 
 ## 3.11.1