+++ packages/widgetbook_cli/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+- **BREAKING**: Set minimum SDK version to 3.9.0. ([#1803](https://github.com/widgetbook/widgetbook/pull/1803))
+- **REFACTOR**: Use `analyzer` 10.x. ([#1804](https://github.com/widgetbook/widgetbook/pull/1804))
+
 ## 3.11.1
 
 - **REFACTOR**: Allow `analyzer` 9.x. ([#1765](https://github.com/widgetbook/widgetbook/pull/1765))